pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ewdiscordbot"
requires-python = ">=3.10"
description = "A stupid bot for fetching quotes and generating markov sentences based on Explorers Wanted characters."
authors = [
  { name = "Daniel Andrlik", email = "daniel@andrlik.org"}
]
license = "BSD-3-Clause"
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
]



dependencies = [
"discord-py-interactions",
"httpx",
"loguru"
]

dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "mypy",
    "mypy-extensions",
    "ipython",
    "types-requests",
    "black",
    "isort",
    "flake8",
    "flake8-isort",
]

[tool.hatch.version]
path = "ewdiscordbot/__about__.py"

[tool.hatch.envs.default]
features = [
  "dev"
]

[tool.hatch.envs.default.scripts]
run_bot = "python ewdiscordbot/run_bot.py"
mtypes = "mypy --install-types --config-file pyproject.toml ./"

[tool.black]
# https://github.com/psf/black
target-version = ["py39"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
    | venvs
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
add_imports = "from __future__ import annotations"
py_version = 39
line_length = 88

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = 3.11
pretty = true
show_traceback = true
color_output = true

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = false
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = true
ignore-nested-functions = false
ignore-nested-classes = true
ignore-setters = false
fail-under = 95
exclude = ["setup.py", "conftest.py", "manage.py", "docs", "build", "tests", "taskhub/migrations", "taskhub/tests"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*", "__str__", "migrations"]
# possible values: 0 (minimal output), 1 (-v), 2 (-vv)
verbose = 2
quiet = false
whitelist-regex = []
color = true
omit-covered-files = false
generate-badge = "."
badge-format = "svg"

[tool.check-wheel-contents]
ignore = ["W004"]
